permisos y autenticacion
se coloco la autentificacion y permisos en casi todas las vistas excepto
las que son para la aplicacin movil

+++ APITX/app/vistas/vistaArticulo.py
@@ -3,45 +3,55 @@ from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from app.models import TxcArticulo
 from app.serializables import TxcArticuloS
+from app.vistas import autentificacion
+from app import permisos
 
 @api_view(['GET', 'POST'])
-def lista_objetos(request):
+def lista_objetos(request, tk):
     """
     Lista de todas las actividades, o crear una nueva
     """
-    if request.method == 'GET':
-        objeto = TxcArticulo.objects.all()
-        serializador = TxcArticuloS(objeto, many=True)
-        return Response(serializador.data)
+    usuario = autentificacion.autenticacion(tk)
+    if usuario.has_perms(permisos.lista_duenios):
+        if request.method == 'GET':
+            objeto = TxcArticulo.objects.all()
+            serializador = TxcArticuloS(objeto, many=True)
+            return Response(serializador.data)
 
-    elif request.method == 'POST':
-        serializador = TxcArticuloS(data=request.data)
-        if serializador.is_valid():
-            serializador.save()
-            return Response(serializador.data,status=status.HTTP_201_CREATED)
-            return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
+        elif request.method == 'POST':
+            serializador = TxcArticuloS(data=request.data)
+            if serializador.is_valid():
+                serializador.save()
+                return Response(serializador.data,status=status.HTTP_201_CREATED)
+                return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
+    else:
+        return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
 
 @api_view(['GET', 'PUT','DELETE'])
-def detalle_objetos(request, pk):
+def detalle_objetos(request, pk, tk):
     """
     Actualiza, elimina un objeto segun su id
     """
-    try:
-        objeto = TxcArticulo.objects.get(pk=pk)
-    except objeto.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    usuario = autentificacion.autenticacion(tk)
+    if usuario.has_perms(permisos.lista_duenios):
+        try:
+            objeto = TxcArticulo.objects.get(pk=pk)
+        except objeto.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
-    if request.method == 'GET':
-        serializador = TxcArticuloS(objeto)
-        return Response(serializador.data)
-
-    elif request.method == 'PUT':
-        serializador = TxcArticuloS(objeto, data=request.data)
-        if serializador.is_valid():
-            serializador.save()
+        if request.method == 'GET':
+            serializador = TxcArticuloS(objeto)
             return Response(serializador.data)
-        return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
 
-    elif request.method == 'DELETE':
-        objeto.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        elif request.method == 'PUT':
+            serializador = TxcArticuloS(objeto, data=request.data)
+            if serializador.is_valid():
+                serializador.save()
+                return Response(serializador.data)
+            return Response(serializador.errors,status=status.HTTP_400_BAD_REQUEST)
+
+        elif request.method == 'DELETE':
+            objeto.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
