+++ APITX/app/vistas/vistaTipodenuncia.py
@@ -5,46 +5,54 @@ from django.core.exceptions import ObjectDoesNotExist
 from app.models import TxdTipodenuncia
 from app.serializables import TxdTipodenunciaS, TxdDenunciaTipoS
 from app import permisos
+from app.vistas import autentificacion
 
 @api_view(['GET', 'POST'])
-def lista_objetos(request):
+def lista_objetos(request, tk):
     """
     Lista de todos los Tipodenuncias, o crea uno nuevo.
     """
-    if request.method == 'GET':
-        objeto = TxdTipodenuncia.objects.all()
-        serializador = TxdDenunciaTipoS(objeto, many=True)
-        return Response(serializador.data)
+    usuario = autentificacion.autenticacion(tk)
+    if usuario.has_perms(permisos.lista_duenios):
+        if request.method == 'GET':
+            objeto = TxdTipodenuncia.objects.all()
+            serializador = TxdDenunciaTipoS(objeto, many=True)
+            return Response(serializador.data)
 
-    elif request.method == 'POST':
-        serializador = TxdTipodenunciaS(data=request.data)
-        if serializador.is_valid():
-            serializador.save()
-            return Response(serializador.data, status=status.HTTP_201_CREATED)
-        return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
-
+        elif request.method == 'POST':
+            serializador = TxdTipodenunciaS(data=request.data)
+            if serializador.is_valid():
+                serializador.save()
+                return Response(serializador.data, status=status.HTTP_201_CREATED)
+            return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
+    else:
+        return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
 
 @api_view(['GET', 'PUT', 'DELETE'])
-def detalle_objetos(request, pk):
+def detalle_objetos(request, pk, tk):
     """
     Actuliza, elimina un objeto segun su id
     """
-    try:
-        objeto = TxdTipodenuncia.objects.get(pk=pk)
-    except ObjectDoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    usuario = autentificacion.autenticacion(tk)
+    if usuario.has_perms(permisos.lista_duenios):
+        try:
+            objeto = TxdTipodenuncia.objects.get(pk=pk)
+        except ObjectDoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
 
-    if request.method == 'GET':
-        serializador = TxdTipodenunciaS(objeto)
-        return Response(serializador.data)
-
-    elif request.method == 'PUT':
-        serializador = TxdTipodenunciaS(objeto, data=request.data)
-        if serializador.is_valid():
-            serializador.save()
+        if request.method == 'GET':
+            serializador = TxdTipodenunciaS(objeto)
             return Response(serializador.data)
-        return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    elif request.method == 'DELETE':
-        objeto.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        elif request.method == 'PUT':
+            serializador = TxdTipodenunciaS(objeto, data=request.data)
+            if serializador.is_valid():
+                serializador.save()
+                return Response(serializador.data)
+            return Response(serializador.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        elif request.method == 'DELETE':
+            objeto.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+    else:
+        return Response("No tiene los permisos necesarios", status=status.HTTP_403_NOT_FOUND)
